feat(backfill): startup-gated stacked-ghost PnL repair
Extracts the backfill logic into a shared service and wires a startup
runner gated by POLYTRADE_RUN_PNL_BACKFILL=1. Setting the env flag
triggers Railway to restart the API; on startup the backfill runs
once before reconciliation begins, logs per-group repair detail, and
proceeds normally. Operation is idempotent so leaving the flag set
is harmless, but recommend unsetting after one successful run.

Admin route (POST /api/admin/backfill-stacked-ghost-pnl) now delegates
to the same shared service for parity. Still requires admin auth.

Repair logic (unchanged):
  - Detect groups of ghost-closed rows in the corruption window that
    share (symbol, side) + exit_time-truncated-to-second + identical
    non-null pnl
  - Rewrite each row's pnl as (rowQty / sumGroupQty) * aggregatePnl
  - Aggregate recovered from the duplicate pnl value

Idempotent: after repair the distinct-pnl detection no longer fires
because pro-rata shares are different values.
